requirements.txt
Flask==0.10.1
Flask-Sockets==0.1
Jinja2==2.7.1
MarkupSafe==0.18
Werkzeug==0.9.4
gevent==0.13.8
gevent-websocket==0.3.6
greenlet==0.4.1
itsdangerous==0.23
websocket-client==0.12.0
wsgiref==0.1.2
